requirements.txt
# Runtime dependencies
pandas
openpyxl
PyYAML
xlrd
psycopg2-binary
python-dotenv
requests
curl_cffi
shapely
selenium
webdriver-manager
pdf2image
pytesseract
openai
anthropic

# Web UI
fastapi
uvicorn[standard]

# SQLAlchemy (used by inventory module and future modules)
sqlalchemy
geoalchemy2
httpx
apscheduler
pydantic-settings

# Local verification
pytest
